src/requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.3
threadpoolctl==3.6.0
Werkzeug==3.1.3
